// frontend/src/app/features/profile/pages/profile/interfaces/profile.interface.ts
export interface Profile {

    _id: string;

    name: string;

    email: string;

    role: string;

    company: string;

    location: string;

    website: string;

    github: string;

    linkedin: string;

    bio: string;

    avatar: string;

    createdAt: string;

    updatedAt: string;

}